requirements.txt
# ----------------------------
# LangChain Ecosystem
# ----------------------------
langchain>=0.3.7,<0.4
langchain-community>=0.3.5,<0.4
langchain-openai>=0.2.8,<0.3
langchain-text-splitters>=0.3.2,<0.4
langchain-experimental>=0.3.3,<0.4

# ----------------------------
# PDF Processing
# ----------------------------
pypdf>=5.1.0
pymupdf>=1.24.13
pdfplumber>=0.11.4

# ----------------------------
# Multi-Format Document Support
# ----------------------------
python-docx>=1.1.0
openpyxl>=3.1.2
xlrd>=2.0.1
pandas>=2.2.3

# ----------------------------
# Text Processing
# ----------------------------
sentence-transformers>=3.3.1
nltk>=3.9.1
spacy>=3.8.2

# ----------------------------
# ML / Embeddings
# ----------------------------
torch>=2.2.0
transformers>=4.45.0

# ----------------------------
# Vector Database
# ----------------------------
qdrant-client>=1.12.1

# ----------------------------
# Retrieval & Ranking
# ----------------------------
rank-bm25>=0.2.2

# ----------------------------
# LLM
# ----------------------------
openai>=1.55.3

# ----------------------------
# Evaluation
# ----------------------------
ragas>=0.2.3
datasets>=3.1.0

# ----------------------------
# Utilities
# ----------------------------
python-dotenv>=1.0.1
pyyaml>=6.0.2
numpy>=1.26.4
pandas>=2.2.3
tqdm>=4.67.1
colorlog>=6.9.0

# ----------------------------
# Web Scraping
# ----------------------------
selenium>=4.15.0
webdriver-manager>=4.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.0

# ----------------------------
# PDF Generation
# ----------------------------
reportlab>=4.0

# ----------------------------
# Web Framework
# ----------------------------
flask>=2.3.0
streamlit>=1.40.0
werkzeug>=2.3.0

# ----------------------------
# Dev
# ----------------------------
jupyter>=1.1.1
ipykernel>=6.29.5
pytest>=8.3.4
black>=24.10.0